郑立231017000175-基于NLP的购物评论情感分析实践/֣��231017000175-����NLP�Ĺ���������з���ʵ��/LSTM_Demo/requirements.txt
gensim (3.7.1)
h5py (2.7.1)
jieba (0.38)
joblib (1.1.1)
Keras (1.1.1)
numpy (1.17.0)
pandas (0.19.1)
pip (9.0.1)
python-dateutil (2.8.2)
pytz (2023.3.post1)
PyYAML (5.4.1)
scikit-learn (0.24.2)
scipy (1.2.1)
setuptools (28.8.0)
six (1.16.0)
smart-open (6.4.0)
Theano (1.0.5)
threadpoolctl (3.1.0)
xlrd (2.0.1)